Create README.md
The Pi Matrix is a generative tool used in ideation, creativity techniques and project completion. The Pi Matrix is presented as a multi-input (4, 8, and 16) algorigram involving an algorithmic execution in three sequential steps. It allows to generate a creative solution according to a multi-tendential context that is intentionally heterogeneous and contradictory.